requirements.txt
pandas
numpy
matplotlib
plotly
statsmodels
scikit-learn
yfinance
streamlit
Pillow

